+++ client/src/pages/openid/openid-list.jsx
@@ -129,13 +129,13 @@ const OpenIdList = () => {
 
   return <div style={{ maxWidth: "1200px", margin: "auto" }}>
     <Stack direction="row" spacing={1} style={{ marginBottom: '20px' }}>
-      <Button variant="contained" color="primary" startIcon={<SyncOutlined />} onClick={() => {
-        refresh();
-      }}>{t('global.refresh')}</Button>&nbsp;&nbsp;
       <Button variant="contained" color="primary" startIcon={<PlusCircleOutlined />} onClick={() => {
         setClientId(null);
         setOpenNewModal(true);
-      }}>{t('global.createAction')}</Button>
+      }}>{t('global.createAction')}</Button>&nbsp;&nbsp;
+      <Button variant="outlined" color="primary" startIcon={<SyncOutlined />} onClick={() => {
+        refresh();
+      }}>{t('global.refresh')}</Button>
     </Stack>
 
     {config && <>