+++ src/utils/utils.go
@@ -887,11 +887,21 @@ func DownloadFileToLocation(path, url string) error {
 }
 
 func GetClientIP(req *http.Request) string {
-	/*ip := req.Header.Get("X-Forwarded-For")
-	if ip == "" {
-		ip = req.RemoteAddr
-	}*/
-	return req.RemoteAddr
+	remoteAddr, _ := SplitIP(req.RemoteAddr)
+
+	if req.Header.Get("x-forwarded-for") != "" && IsTrustedProxy(remoteAddr) {
+		remoteAddr, _ = SplitIP(strings.TrimSpace(strings.Split(req.Header.Get("X-Forwarded-For"), ",")[0]))
+	}
+	return remoteAddr
+}
+
+func IsTrustedProxy(ip string) bool {
+	for _, trustedProxy := range GetMainConfig().HTTPConfig.TrustedProxies {
+		if isInRange, _ := IPInRange(ip, trustedProxy); isInRange {
+			return true
+		}
+	}
+	return false
 }
 
 func IsDomain(domain string) bool {