+++ changelog.md
@@ -4,6 +4,7 @@
  - Added "skip clean URL" option for apps that have invalid URLs like Synology
  - Support for tempFS
  - Improve support for 0.0.0.0 routes
+ - Added trusted proxy option (Thanks @InterN0te)
  - Reduce SmartShield false positive on the server panel by having two level of strictness (UI/panel vs. login)
  - Rclone rework, no more sub processes (increased performance and reliability)
  - Reworked Constellation cluster synchronisation completely. Now there are no more "Master" server and each server is equally capable.